requirements.txt
asgiref==3.2.10
certifi==2020.6.20
cffi==1.14.1
chardet==3.0.4
cloudinary==1.23.0
cryptography==3.0
dj-database-url==0.5.0
dj3-cloudinary-storage==0.0.3
Django==3.1
django-braces==1.14.0
django-cloudmailin==0.1.1
django-cors-headers==3.5.0
django-countries==6.1.2
django-crispy-forms==1.9.2
django-environ==0.4.5
django-heroku==0.3.1
django-ranged-response==0.2.0
django-simple-captcha==0.5.13
djangorestframework==3.12.2
et-xmlfile==1.0.1
gunicorn==20.0.4
idna==2.10
jdcal==1.4.1
oauthlib==3.1.0
openpyxl==3.0.5
Pillow==8.2.0
psycopg2==2.8.5
pycparser==2.20
PyJWT==1.7.1
python3-openid==3.2.0
pytz==2020.1
requests==2.24.0
requests-oauthlib==1.3.0
six==1.15.0
social-auth-app-django==4.0.0
social-auth-core==3.3.3
sqlparse==0.3.1
urllib3==1.25.10
whitenoise==5.2.0
git+git://github.com/AndreaTuci/django-highcharts.git
